+++ src/Native/CollectionTrait.php
@@ -11,20 +11,19 @@ namespace DecodeLabs\Collections\Native;
 
 use ArrayIterator;
 use Closure;
-
 use DecodeLabs\Coercion;
 use DecodeLabs\Collections\ArrayUtils;
 use DecodeLabs\Collections\Collection;
 use DecodeLabs\Collections\Sequence;
 use DecodeLabs\Exceptional;
 use DecodeLabs\Fluidity\ThenTrait;
-
 use Iterator;
 use ReflectionFunction;
 
 /**
  * @template TKey
  * @template TValue
+ * @phpstan-require-implements Collection<TKey, TValue>
  */
 trait CollectionTrait
 {
@@ -88,7 +87,7 @@ trait CollectionTrait
      * Get first item, matching filter
      */
     public function getFirst(
-        callable $filter = null
+        ?callable $filter = null
     ): mixed {
         /* @phpstan-ignore-next-line */
         return ArrayUtils::getFirst($this->items, $filter, $this);
@@ -98,7 +97,7 @@ trait CollectionTrait
      * Get the last item in the list, matching filter
      */
     public function getLast(
-        callable $filter = null
+        ?callable $filter = null
     ): mixed {
         /* @phpstan-ignore-next-line */
         return ArrayUtils::getLast($this->items, $filter, $this);
@@ -217,8 +216,9 @@ trait CollectionTrait
      */
     public function slice(
         int $offset,
-        int $length = null
+        ?int $length = null
     ): static {
+        // @phpstan-ignore-next-line
         return $this->propagate(array_slice(
             $this->items,
             $offset,
@@ -233,6 +233,7 @@ trait CollectionTrait
     public function sliceRandom(
         int $number
     ): static {
+        // @phpstan-ignore-next-line
         return $this->propagate(ArrayUtils::sliceRandom($this->items, $number));
     }
 
@@ -251,6 +252,7 @@ trait CollectionTrait
             $output[] = $this->propagate($chunk);
         }
 
+        // @phpstan-ignore-next-line
         return $output;
     }
 
@@ -268,6 +270,7 @@ trait CollectionTrait
             $output[] = $this->propagate($chunk);
         }
 
+        // @phpstan-ignore-next-line
         return $output;
     }
 
@@ -291,6 +294,7 @@ trait CollectionTrait
     public function diffAssoc(
         iterable ...$arrays
     ): static {
+        // @phpstan-ignore-next-line
         return $this->propagate(array_diff_assoc(
             $this->items,
             ...ArrayUtils::iterablesToArrays(...$arrays)
@@ -312,6 +316,7 @@ trait CollectionTrait
 
         $args[] = $keyCallback;
 
+        // @phpstan-ignore-next-line
         return $this->propagate(
             /** @phpstan-ignore-next-line */
             array_diff_uassoc($this->items, ...$args)
@@ -333,6 +338,7 @@ trait CollectionTrait
 
         $args[] = $valueCallback;
 
+        // @phpstan-ignore-next-line
         return $this->propagate(
             /** @phpstan-ignore-next-line */
             array_udiff_assoc($this->items, ...$args)
@@ -356,6 +362,7 @@ trait CollectionTrait
         $args[] = $valueCallback;
         $args[] = $keyCallback;
 
+        // @phpstan-ignore-next-line
         return $this->propagate(
             /** @phpstan-ignore-next-line */
             array_udiff_uassoc($this->items, ...$args)
@@ -368,6 +375,7 @@ trait CollectionTrait
     public function diffValues(
         iterable ...$arrays
     ): static {
+        // @phpstan-ignore-next-line
         return $this->propagate(array_diff(
             $this->items,
             ...ArrayUtils::iterablesToArrays(...$arrays)
@@ -389,6 +397,7 @@ trait CollectionTrait
 
         $args[] = $valueCallback;
 
+        // @phpstan-ignore-next-line
         return $this->propagate(
             /** @phpstan-ignore-next-line */
             array_udiff($this->items, ...$args)
@@ -401,6 +410,7 @@ trait CollectionTrait
     public function diffKeys(
         iterable ...$arrays
     ): static {
+        // @phpstan-ignore-next-line
         return $this->propagate(array_diff_key(
             $this->items,
             ...ArrayUtils::iterablesToArrays(...$arrays)
@@ -422,6 +432,7 @@ trait CollectionTrait
 
         $args[] = $keyCallback;
 
+        // @phpstan-ignore-next-line
         return $this->propagate(
             /** @phpstan-ignore-next-line */
             array_diff_ukey($this->items, ...$args)
@@ -435,6 +446,7 @@ trait CollectionTrait
     public function intersectAssoc(
         iterable ...$arrays
     ): static {
+        // @phpstan-ignore-next-line
         return $this->propagate(array_intersect_assoc(
             $this->items,
             ...ArrayUtils::iterablesToArrays(...$arrays)
@@ -456,6 +468,7 @@ trait CollectionTrait
 
         $args[] = $keyCallback;
 
+        // @phpstan-ignore-next-line
         return $this->propagate(
             /** @phpstan-ignore-next-line */
             array_intersect_uassoc($this->items, ...$args)
@@ -477,6 +490,7 @@ trait CollectionTrait
 
         $args[] = $valueCallback;
 
+        // @phpstan-ignore-next-line
         return $this->propagate(
             /** @phpstan-ignore-next-line */
             array_uintersect_assoc($this->items, ...$args)
@@ -500,6 +514,7 @@ trait CollectionTrait
         $args[] = $valueCallback;
         $args[] = $keyCallback;
 
+        // @phpstan-ignore-next-line
         return $this->propagate(
             /** @phpstan-ignore-next-line */
             array_uintersect_uassoc($this->items, ...$args)
@@ -511,6 +526,7 @@ trait CollectionTrait
      */
     public function intersectValues(iterable ...$arrays): static
     {
+        // @phpstan-ignore-next-line
         return $this->propagate(array_intersect(
             $this->items,
             ...ArrayUtils::iterablesToArrays(...$arrays)
@@ -532,6 +548,7 @@ trait CollectionTrait
 
         $args[] = $valueCallback;
 
+        // @phpstan-ignore-next-line
         return $this->propagate(
             /** @phpstan-ignore-next-line */
             array_uintersect($this->items, ...$args)
@@ -543,6 +560,7 @@ trait CollectionTrait
      */
     public function intersectKeys(iterable ...$arrays): static
     {
+        // @phpstan-ignore-next-line
         return $this->propagate(array_intersect_key(
             $this->items,
             ...ArrayUtils::iterablesToArrays(...$arrays)
@@ -564,6 +582,7 @@ trait CollectionTrait
 
         $args[] = $keyCallback;
 
+        // @phpstan-ignore-next-line
         return $this->propagate(
             /** @phpstan-ignore-next-line */
             array_intersect_ukey($this->items, ...$args)
@@ -575,11 +594,13 @@ trait CollectionTrait
      * Return subset of collection where callback returns true
      */
     public function filter(
-        callable $callback = null
+        ?callable $callback = null
     ): static {
         if ($callback) {
+            // @phpstan-ignore-next-line
             return $this->propagate(array_filter($this->items, $callback, ARRAY_FILTER_USE_BOTH));
         } else {
+            // @phpstan-ignore-next-line
             return $this->propagate(array_filter($this->items));
         }
     }
@@ -591,6 +612,7 @@ trait CollectionTrait
         callable $callback,
         iterable ...$arrays
     ): static {
+        // @phpstan-ignore-next-line
         return $this->propagate(array_map(
             $callback,
             $this->items,
@@ -614,6 +636,7 @@ trait CollectionTrait
                 $output = array_merge($output, iterator_to_array($callback($value, $key)));
             }
 
+            // @phpstan-ignore-next-line
             return $this->propagate($output);
         } else {
             $keys = array_keys($this->items);
@@ -623,6 +646,7 @@ trait CollectionTrait
                 throw Exceptional::UnexpectedValue('Combine failed - key count does not match item count');
             }
 
+            // @phpstan-ignore-next-line
             return $this->propagate((array)array_combine($keys, $items));
         }
     }
@@ -642,7 +666,7 @@ trait CollectionTrait
      * Add all numeric values in collection
      */
     public function getSum(
-        callable $filter = null
+        ?callable $filter = null
     ): float {
         return Coercion::toFloat(
             $this->reduce(function ($result, $item) use ($filter) {
@@ -663,7 +687,7 @@ trait CollectionTrait
      * Multiple all numeric values in collection
      */
     public function getProduct(
-        callable $filter = null
+        ?callable $filter = null
     ): float {
         return Coercion::toFloat(
             $this->reduce(function ($result, $item) use ($filter) {
@@ -684,7 +708,7 @@ trait CollectionTrait
      * Get average value of numerics
      */
     public function getAvg(
-        callable $filter = null
+        ?callable $filter = null
     ): ?float {
         if (!$count = count($this->items)) {
             return null;
@@ -699,7 +723,7 @@ trait CollectionTrait
      */
     public function pluck(
         string $valueKey,
-        string $indexKey = null
+        ?string $indexKey = null
     ): array {
         return array_column($this->items, $valueKey, $indexKey);
     }
@@ -770,7 +794,7 @@ trait CollectionTrait
     /**
      * Convert to array
      *
-     * @return array<TKey, TValue>
+     * @return array<TKey,TValue>
      */
     public function toArray(): array
     {
@@ -780,7 +804,7 @@ trait CollectionTrait
     /**
      * Convert to json
      *
-     * @return array<int|string, mixed>
+     * @return array<int|string,mixed>
      */
     public function jsonSerialize(): array
     {
@@ -792,7 +816,7 @@ trait CollectionTrait
     /**
      * Get dump info
      *
-     * @return array<TKey, TValue>
+     * @return array<TKey,TValue>
      */
     public function __debugInfo(): array
     {
@@ -803,7 +827,7 @@ trait CollectionTrait
     /**
      * Copy and reinitialise new object
      *
-     * @return static<TKey, TValue>
+     * @return static<TKey,TValue>
      */
     abstract protected static function propagate(
         iterable $newItems = []